Moved gamescene handlers to a separate folder
