+++ StortSpelProjekt/Game/src/main.cpp
@@ -12,8 +12,8 @@
 #include "Shop.h"
 #include "Components/CurrencyComponent.h"
 #include "Components/UpgradeComponents/UpgradeComponent.h"
-#include "MainMenuHandler.h"
-#include "GameOverHandler.h"
+#include "GameSceneHandlers/MainMenuHandler.h"
+#include "GameSceneHandlers/GameOverHandler.h"
 #include "PauseGUI.h"
 #include "GameTracker.h"
 #include "EnemyStatDefine.h"
